LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-createHTMLDocument-expected.txt

CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedScript value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
shouldpass [default]

PASS Trusted Type assignment is blocked. (document)
PASS Trusted Type instances created in the main doc can be used. (document)
PASS Trusted Type assignment is blocked. (createHTMLDocument)
PASS Trusted Type instances created in the main doc can be used. (createHTMLDocument)
PASS Trusted Type assignment is blocked. (DOMParser)
PASS Trusted Type instances created in the main doc can be used. (DOMParser)
PASS Trusted Type assignment is blocked. (XHR)
PASS Trusted Type instances created in the main doc can be used. (XHR)
PASS Install default policy.
PASS Default policy applies. (document)
PASS Default policy applies. (createHTMLDocument)
PASS Default policy applies. (DOMParser)
PASS Default policy applies. (XHR)
